+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "bdsp-evscript-language-support" extension will be do
 ## [1.3]
 
 - Command Palette commands for parsing and assembling ev_scripts using ev-as
+- Configuration Options to enable message validation during ev-as assembly
 
 ## [1.2]